+++ scripts/macros/2014/classFeatures/monk/ki.js
@@ -1,4 +1,5 @@
 export let ki = {
     name: 'Ki Points',
+    aliases: ['Ki'],
     version: '1.1.0'
 };