+++ CHANGELOG.md
@@ -12,6 +12,11 @@
 ## Bug Fixes
 - Chromatic Orb (Modern)
 - Unconscious condition
+- Summon Fey
+- Embedded Macros
+- Spirit Shroud
+- Relentless Endurance (Legacy)
+- Item Medkit
 
 # 1.2.40 Release:
 ## Bug Fixes